tag: fix inference in generic scopes
