+++ kyo-tag/shared/src/test/scala/kyo/internal/IntEncoderTest.scala
